requirements.txt
aioredis==1.3.1
amqp==2.5.2
appdirs==1.4.4
asgi-redis==1.4.3
asgiref==3.2.7
async-timeout==3.0.1
asyncio==3.4.3
attrs==19.3.0
autobahn==20.4.2
Automat==20.2.0
billiard==3.6.3.0
celery==4.4.2
cffi==1.14.0
channels==2.4.0
channels-redis==2.4.2
constantly==15.1.0
cryptography==2.9.2
daphne==2.5.0
distlib==0.3.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.0.5
django-bootstrap-form==3.4
django-braces==1.14.0
django-categories==1.7.1
django-category==2.0.1
django-celery-beat==2.0.0
django-crispy-forms==1.9.0
django-debug-toolbar==2.2
django-favicon==0.1.3
django-friendship==1.9.1
django-js-asset==1.2.2
django-league==0.1.1
django-mathfilters==1.0.0
django-messages==0.6.0
django-model-utils==4.0.0
django-mptt==0.9.0
django-notifications-hq==1.6.0
django-object-tools==2.0.0
django-private-chat==0.2.0
django-star-ratings==0.9.1
django-timezone-field==4.0
django-widget-tweaks==1.4.8
djangorestframework==3.11.0
dnspython==1.16.0
eventlet==0.25.2
filelock==3.0.12
greenlet==0.4.15
hiredis==1.0.1
hyperlink==19.0.0
idna==2.9
importlib-metadata==1.6.1
incremental==17.5.0
jsonfield==3.1.0
kombu==4.6.8
monotonic==1.5
msgpack==0.6.0
msgpack-python==0.5.6
mysqlclient==1.4.6
Pillow==7.1.1
pycparser==2.20
PyHamcrest==2.0.2
pystache==0.5.4
python-crontab==2.4.1
python-dateutil==2.8.1
pytz==2019.3
redis==3.4.1
six==1.14.0
South==1.0.2
sqlparse==0.3.1
static==1.1.1
static3==0.7.0
stream-framework==1.4.0
swapper==1.1.2
Twisted==20.3.0
txaio==20.4.1
unicode-slugify==0.1.3
Unidecode==1.1.1
utils==1.0.1
vine==1.3.0
virtualenv==20.0.18
websockets==8.1
zipp==3.1.0
zope.interface==5.1.0
